token 로직 새로 짜야할것 같음. ai 믿지말자...

+++ src/main/java/com/xodud1202/springbackend/security/JwtTokenProvider.java
@@ -17,10 +17,10 @@ public class JwtTokenProvider {
     private String jwtSecret;
     
     @Value("${jwt.access-token-expiration}")
-    private int jwtAccessTokenExpirationInMs;
+    private long jwtAccessTokenExpirationInMs;
     
     @Value("${jwt.refresh-token-expiration}")
-    private int jwtRefreshTokenExpirationInMs;
+    private long jwtRefreshTokenExpirationInMs;
     
     // Access Token 생성
     public String generateAccessToken(Authentication authentication) {
